+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## least.js 2 Changelog
-> last change: 16/01/2014 - initial release
+> last change: 21/01/2014 - added caption element for "image-preview"
+
+## V - 2.1 (20/01/2014)
+> added caption element for "image-preview"
 
 ## V - 2.0.1 (16/01/2014)
 > initial release